refactor(server): drop redundant cloud prefixes from azure wire dirs (#325)
