+++ server/azure/ai/accounts.go
@@ -1,4 +1,4 @@
-package azureai
+package ai
 
 import (
 	"net/http"
