+++ server/azure/ai/dataplane.go
@@ -1,4 +1,4 @@
-package azureai
+package ai
 
 import (
 	"encoding/json"
