+++ server/azure/ai/machinelearning_children.go
@@ -1,4 +1,4 @@
-package azureai
+package ai
 
 import (
 	"net/http"
