+++ server/azure/ai/machinelearning_simple.go
@@ -1,4 +1,4 @@
-package azureai
+package ai
 
 import (
 	"net/http"
