+++ server/azure/sql/operations.go
@@ -1,4 +1,4 @@
-package azuresql
+package sql
 
 import (
 	"context"
